requirements.txt
django~=1.11
djangorestframework
drf-nested-routers
django-dotenv
mysqlclient
coinbase
